backend/users/urls.py
# core/urls.py
from rest_framework.routers import DefaultRouter
from .views import UserViewSet

router = DefaultRouter()
router.register(r'', UserViewSet, basename='user')  # /backend/users/

urlpatterns = router.urls
