backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
Pillow==11.2.1
python-decouple==3.8
